Move parser to another file
