+++ src/const.hpp
@@ -1,0 +1,6 @@
+#define ENV_NAME "DEBUG"
+#define BUF_SIZE 1024
+#define FD_OUT 3
+#define FD_ERR 4
+#define BEGIN_OUT "B_OUT"
+#define END_OUT "E_OUT"
