+++ src/main.cpp
@@ -1,32 +1,20 @@
 #include <iostream>
-#include <cstdlib>
 #include <unistd.h>
 #include <cstring>
 #include <fcntl.h>
 #include <chrono>
-#include <fstream>
-#include <string>
 #include <iomanip>
-#include <map>
+
+#include "parser.hpp"
+#include "const.hpp"
 
 using namespace std;
 
-#define ENV_NAME "DEBUG"
-#define BUF_SIZE 1024
-#define FD_OUT 3
-#define FD_ERR 4
-#define BEGIN_OUT "B_OUT"
-#define END_OUT "E_OUT"
-
-string program_name;
-
-void writePackage(char *, bool);
-void writeHeader(ofstream &, string);
-
-map<string, ofstream> pids;
-
 int rootProcess()
 {
+    map<string, ofstream> pids;
+    pids.clear();
+
     // SELECT
     fd_set read_fds;
     int select_fd = 5;
@@ -53,7 +41,7 @@ int rootProcess()
             if (rbytes <= 0)
                 out_ended = true;
             else
-                writePackage(buf, false);
+                writePackage(pids, buf, false);
         }
 
         if (!err_ended && FD_ISSET(FD_ERR, &read_fds))
@@ -63,9 +51,12 @@ int rootProcess()
             if (rbytes <= 0)
                 err_ended = true;
             else
-                writePackage(buf, true);
+                writePackage(pids, buf, true);
         }
     }
+
+    for (auto &[k, v] : pids)
+        v.close();
     return 0;
 }
 
@@ -150,89 +141,6 @@ int childProcess(char *program, char *argv[])
         }
     }
     return 0;
-}
-
-void writeHeader(ofstream &result, string program_name)
-{
-    auto now = chrono::system_clock::now();
-    auto time = chrono::system_clock::to_time_t(now);
-    result << R"(<!DOCTYPE html>
-<html lang="pl-PL">
-<head>
-<meta charset="UTF-8" />
-<link rel="stylesheet" href="styles.css" />
-<title>)";
-    result << program_name << "\n";
-    result << R"(</title>
-</head>
-<body>
-<div class="head">
-<div class="info">
-<span class="info-title">command:</span> <span class="info-value info-value-path">)" +
-                  program_name + R"(</span>
-</div>
-<div class="info">
-<span class="info-title">start time:</span> <span class="info-value">)";
-    result << put_time(localtime(&time), "%Y-%m-%d %H:%M:%S");
-    result << R"(</span>
-</div>
-<div class="info">
-<span class="info-title">last entry:</span> <span class="info-value"></span>
-</div>
-</div>
-<div class="line"></div>
-<div class="entries">
-<table>
-<tboby>
-)";
-    result.flush();
-}
-
-void writeLine(ofstream &result, string line, string timeStr, bool isError)
-{
-    long long milliseconds = stoll(timeStr);
-    auto duration = chrono::milliseconds(milliseconds);
-    auto time_point = chrono::system_clock::time_point(duration);
-    time_t time = chrono::system_clock::to_time_t(time_point);
-
-    result << "<tr><td class=\"entry-time\">";
-    result << put_time(localtime(&time), "%Y-%m-%d %H:%M:%S");
-    if (isError)
-        result << "</td><td>&nbsp;</td><td class=\"entry-log entry-stderr\">";
-    else
-        result << "</td><td>&nbsp;</td><td class=\"entry-log entry-stdout\">";
-    result << line;
-    result << "</td></tr>\n";
-    result.flush();
-}
-
-void writePackage(char *buf, bool isError)
-{
-    istringstream stream(buf);
-    string line;
-    string pid, time;
-
-    while (getline(stream, line))
-    {
-        if (line[0] == 'B')
-        {
-            getline(stream, pid);
-            getline(stream, time);
-            continue;
-        }
-        if (line[0] == 'E')
-        {
-            continue;
-        }
-
-        if (pids.find(pid) == pids.end())
-        {
-            pids[pid].open("result_" + pid + ".html", ios::out | ios::trunc);
-            writeHeader(pids[pid], "");
-        }
-        ofstream &s = pids[pid];
-        writeLine(s, line, time, isError);
-    }
 }
 
 int main(int, char *argv[])
